+++ map.h
@@ -230,7 +230,7 @@ namespace Maps {
             zone.emplace<Component::Entity_Type>(entity, Component::Entity_Type::unit);
             zone.emplace<Component::Action>(entity, Component::idle);
             zone.emplace<Component::Velocity>(entity, 0.0f, 0.0f, 0.0f, 0.0f, data.speed * data.scale);
-            auto &health = zone.emplace<Component::Health>(entity, data.health * data.scale);
+            auto &health = zone.emplace<Component::Health>(entity, int(data.health * data.scale));
             zone.emplace<Component::Soldier>(entity);
             zone.emplace<Component::Commandable>(entity);
             zone.emplace<Component::Spellbook>(entity);
